fxies

+++ Crypto/Domain/Entities/Transaction.cs
@@ -12,7 +12,7 @@ namespace Binebase.Exchange.CryptoService.Domain.Entities
         public ulong RawAmount { get; set; }
         public decimal Amount { get; set; }
         public DateTime? Confirmed { get; set; }
-        public int Confirmations { get; set; }
+        public ulong Confirmations { get; set; }
         public TransactionStatus Status { get; set; }
         public Guid AddressId { get; set; }
         public Address Address { get; set; }
