+++ Crypto/Infrastructure/Configuration.cs
@@ -9,7 +9,7 @@ namespace Binebase.Exchange.CryptoService.Infrastructure
         public string Mnemonic { get; set; }
         public string Password { get; set; }
         public int WithdrawAccountIndex { get; set; }
-        public int ConfirmationsCount { get; set; }
+        public ulong ConfirmationsCount { get; set; }
         public TimeSpan TransactionPoolingTimeout { get; set; }
         public string EtherscanUrlFormat { get; set; }
         public Uri EthereumNode { get; set; }
